Implement frequency model
The methods fit and predict of the predictor were entangled in the
pipeline, which make the model hard to use and modify and the pipeline
hard to read and understand.

For convenience, this change introduces a model for the
frequency predictor (with methods fit and predict). This will allow
the cross validation script to run more easily. This also draws an
inspiring format for our future models.

[Close #3]