Implement the submissioner
Its purpose is to make the creation of a new submit file easier.

The function is : create_submission(mids_prediction, output_filename=None)
mids_prediction is of the form : {mid: prediction}
where prediction is a list of email addresses.

When no filename is given, the file is saved to the submissions folder
and contains the date in its name.

[Close #5]