Change the format of argument in score function
In metrics.mean_average_precision, the arguments were first a list
of lists. We decided to change the format of prediction to dict of
lists for convenience.

This change adapts metrics.mean_average_precision,
in order to be able to compute a score from the output of model.predict
without any parsing