Implement cross-validation script.
The script "cross_validation_script.py" does a random k-fold cross validation.

The number of splits, the random seed and the model have to be changed directly
in the script.
We also add the function true_recipients(y_df) in general_tools/format_data.py
The pipeline script now computes the training score.

[Close #2]