Take into account co-occurences
In the LinagoraWinningPredictor, the output is now structured
according to empirical co-occurences probabilities.

However, for the moment it takes too much time to really be usable
in practice (and to be cross validated). I add it to master anyway,
and will improve the execution time later. For the moment it takes
1 hour to compute the prediction.

[Close #15]