Add persistent memoization to co occurrences
This allows to store in a file (at exit) the computation of
co occurences. This means the first time you compute it, it will take
a long time to run, but it will be much faster the second time (even
if you KeyboardInterrupt the execution).

Be careful to use it correctly (if the features change, or the X_train
changes, the precomputed values will be wrong)